Update Project2.txt
For last commit: Created strings for a long input, an invalid input, and an empty input. Set aside space for 8 bits. Added Error for a long input string; copy RAM address of str0 and load it into $a0. Load 4 immediate to $v0 register. Jumps to program's exit.

For this commit: Added Error for empty and invalid inputs; repeat process for each string.